Automatically go to the starter pack booster if it's the first time visiting
This is a temporary change I'm making in anticipation of the
steam demo. Probably there will be a lot of new people, so
I'd like them to get the best possible impression of custom
levels from the beginning.

I'll leave it like this for a week or so then revert probably